Updated the facebook login for version 3.2 of the API
v3.2 of the FAcebook login API requires that developers explicityl specify the email parameter on the request in order to see the user's email int he result of the request.  So I just added it to the request parameter as required and now the API can reutnr user emails. It might also be useful to request for the phone number in case the user created the account using a phone number.